auth (kernel v9.1)

Erlang network authentication server.

This module is deprecated. For a description of the Magic Cookie system, refer to Distributed Erlang in the Erlang Reference Manual.

Types

calls/0

-type calls() :: echo | sync_cookie | {get_cookie, node()} | {set_cookie, node(), term()}.

cookie()

-type cookie() :: atom().

state/0

-type state() :: #state{}.

Functions

cookie()

-spec cookie() -> Cookie when Cookie :: cookie().

Use erlang:get_cookie() in ERTS instead.

cookie/1

-spec cookie(TheCookie) -> true when TheCookie :: Cookie | [Cookie], Cookie :: cookie().

Use erlang:set_cookie(node(), Cookie) in ERTS instead.

is_auth(Node)

-spec is_auth(Node) -> yes | no when Node :: node().

Returns yes if communication with Node is authorized. Notice that a connection to Node is established in this case. Returns no if Node does not exist or communication is not authorized (it has another cookie than auth thinks it has).

Use net_adm:ping(Node) instead.

node_cookie/1

-spec node_cookie(Cookies :: [node() | cookie(), ...]) -> yes | no.

node_cookie([Node, Cookie]) -> yes | no

Equivalent to node_cookie(Node, Cookie).

node_cookie(Node, Cookie)

-spec node_cookie(Node, Cookie) -> yes | no when Node :: node(), Cookie :: cookie().

Sets the magic cookie of Node to Cookie and verifies the status of the authorization. Equivalent to calling erlang:set_cookie(Node, Cookie), followed by auth:is_auth(Node).