auth (kernel v9.3)

This module is deprecated. See each function for what to use instead.

Erlang network authentication server.

For a description of the Magic Cookie system, refer to Distributed Erlang in the Erlang Reference Manual.

Types

cookie()
(not exported)

-type cookie() :: atom().

Functions

cookie()

This function is deprecated. auth:cookie/0 is deprecated; use erlang:get_cookie/0 instead.

-spec cookie() -> Cookie when Cookie :: cookie().

Use erlang:get_cookie() in ERTS instead.

cookie(TheCookie)

This function is deprecated. auth:cookie/1 is deprecated; use erlang:set_cookie/2 instead.

-spec cookie(TheCookie) -> true when TheCookie :: Cookie | [Cookie], Cookie :: cookie().

Use erlang:set_cookie(node(), Cookie) in ERTS instead.

is_auth(Node)

This function is deprecated. auth:is_auth/1 is deprecated; use net_adm:ping/1 instead.

-spec is_auth(Node) -> yes | no when Node :: node().

Returns yes if communication with Node is authorized. Notice that a connection to Node is established in this case. Returns no if Node does not exist or communication is not authorized (it has another cookie than auth thinks it has).

Use net_adm:ping(Node) instead.

node_cookie(Cookies)

This function is deprecated. auth:node_cookie/1 is deprecated; use erlang:set_cookie/2 and net_adm:ping/1 instead.

-spec node_cookie(Cookies :: [node() | cookie(), ...]) -> yes | no.

node_cookie([Node, Cookie]) -> yes | no

Equivalent to node_cookie(Node, Cookie).

node_cookie(Node, Cookie)

This function is deprecated. auth:node_cookie/2 is deprecated; use erlang:set_cookie/2 and net_adm:ping/1 instead.

-spec node_cookie(Node, Cookie) -> yes | no when Node :: node(), Cookie :: cookie().

Sets the magic cookie of Node to Cookie and verifies the status of the authorization. Equivalent to calling erlang:set_cookie(Node, Cookie), followed by auth:is_auth(Node).