logger (kernel v8.0.2)

This module implements the main API for logging in Erlang/OTP. To create a log event, use the API functions or the log macros, for example:

?LOG_ERROR("error happened because: ~p", [Reason]).   % With macro
logger:error("error happened because: ~p", [Reason]). % Without macro
    

To configure the Logger backend, use Kernel configuration parameters or configuration functions in the Logger API.

By default, the Kernel application installs one log handler at system start. This handler is named default. It receives and processes standard log events produced by the Erlang runtime system, standard behaviours and different Erlang/OTP applications. The log events are by default printed to the terminal.

If you want your systems logs to be printed to a file instead, you must configure the default handler to do so. The simplest way is to include the following in your sys.config:

[{kernel,
  [{logger,
    [{handler, default, logger_std_h,
      #{config => #{file => "path/to/file.log"}}}]}]}].
      

For more information about:

the Logger facility in general, see the User's Guide.

how to configure Logger, see the Configuration section in the User's Guide.

the built-in handlers, see logger_std_h and logger_disk_log_h.

the built-in formatter, see logger_formatter.

built-in filters, see logger_filters.

Since Logger is new in Erlang/OTP 21.0, we do reserve the right to introduce changes to the Logger API and functionality in patches following this release. These changes might or might not be backwards compatible with the initial version.

Macros

The following macros are defined in logger.hrl, which is included in a module with the directive

    -include_lib("kernel/include/logger.hrl").

?LOG_EMERGENCY(StringOrReport[,Metadata])

?LOG_EMERGENCY(FunOrFormat,Args[,Metadata])

?LOG_ALERT(StringOrReport[,Metadata])

?LOG_ALERT(FunOrFormat,Args[,Metadata])

?LOG_CRITICAL(StringOrReport[,Metadata])

?LOG_CRITICAL(FunOrFormat,Args[,Metadata])

?LOG_ERROR(StringOrReport[,Metadata])

?LOG_ERROR(FunOrFormat,Args[,Metadata])

?LOG_WARNING(StringOrReport[,Metadata])

?LOG_WARNING(FunOrFormat,Args[,Metadata])

?LOG_NOTICE(StringOrReport[,Metadata])

?LOG_NOTICE(FunOrFormat,Args[,Metadata])

?LOG_INFO(StringOrReport[,Metadata])

?LOG_INFO(FunOrFormat,Args[,Metadata])

?LOG_DEBUG(StringOrReport[,Metadata])

?LOG_DEBUG(FunOrFormat,Args[,Metadata])

?LOG(Level,StringOrReport[,Metadata])

?LOG(Level,FunOrFormat,Args[,Metadata])

All macros expand to a call to Logger, where Level is taken from the macro name, or from the first argument in the case of the ?LOG macro. Location data is added to the metadata as described under the metadata() type definition.

The call is wrapped in a case statement and will be evaluated only if Level is equal to or below the configured log level.

See Also

config(4), erlang(3), io(3), logger_disk_log_h(3), logger_filters(3), logger_formatter(3), logger_std_h(3), unicode(3)

Types

-type filter() :: term().

Specs

filter() :: {fun((log_event(), filter_arg()) -> filter_return()), filter_arg()}.

A filter which can be installed as a handler filter, or as a primary filter in Logger.

-type filter_arg() :: term().

Specs

filter_arg() :: term().

The second argument to the filter fun.

-type filter_id() :: term().

Specs

filter_id() :: atom().

A unique identifier for a filter.

-type filter_return() :: term().

Specs

filter_return() :: stop | ignore | log_event().

The return value from the filter fun.

-type formatter_config() :: term().

Specs

formatter_config() :: #{atom() => term()}.

Configuration data for the formatter. See logger_formatter(3) for an example of a formatter implementation.

-type handler_config() :: term().

Specs

handler_config() ::
    #{id => handler_id(),
      config => term(),
      level => level() | all | none,
      module => module(),
      filter_default => log | stop,
      filters => [{filter_id(), filter()}],
      formatter => {module(), formatter_config()}}.

Handler configuration data for Logger. The following default values apply:

level => all

filter_default => log

filters => []

formatter => {logger_formatter, DefaultFormatterConfig}

In addition to these, the following fields are automatically inserted by Logger, values taken from the two first parameters to add_handler/3:

id => HandlerId

module => Module

These are read-only and cannot be changed in runtime.

Handler specific configuration data is inserted by the handler callback itself, in a sub structure associated with the field named config. See the logger_std_h(3) and logger_disk_log_h(3) manual pages for information about the specific configuration for these handlers.

See the logger_formatter(3) manual page for information about the default configuration for this formatter.

-type handler_id() :: term().

Specs

handler_id() :: atom().

A unique identifier for a handler instance.

-type level() :: term().

Specs

level() :: emergency | alert | critical | error | warning | notice | info | debug.

The severity level for the message to be logged.

-type log_event() :: term().

Specs

log_event() ::
    #{level := level(),
      msg := {io:format(), [term()]} | {report, report()} | {string, unicode:chardata()},
      meta := metadata()}.

-type metadata() :: term().

Specs

metadata() ::
    #{pid => pid(),
      gl => pid(),
      time => timestamp(),
      mfa => {module(), atom(), non_neg_integer()},
      file => file:filename(),
      line => non_neg_integer(),
      domain => [atom()],
      report_cb => report_cb(),
      atom() => term()}.

Metadata for the log event.

Logger adds the following metadata to each log event:

pid => self()

gl => group_leader()

time => logger:timestamp()

When a log macro is used, Logger also inserts location information:

mfa => {?MODULE, ?FUNCTION_NAME, ?FUNCTION_ARITY}

file => ?FILE

line => ?LINE

You can add custom metadata, either by:

specifying a map as the last parameter to any of the log macros or the logger API functions.

setting process metadata with set_process_metadata/1 or update_process_metadata/1.

setting primary metadata with set_primary_config/1 or through the kernel configuration parameter logger_metadata

When adding custom metadata, make sure not to use any of the keys mentioned above as that may cause a lot of confusion about the log events.

Logger merges all the metadata maps before forwarding the log event to the handlers. If the same keys occur, values from the log call overwrite process metadata, which overwrites the primary metadata, which in turn overwrite values set by Logger.

The following custom metadata keys have special meaning:

domain

The value associated with this key is used by filters for grouping log events originating from, for example, specific functional areas. See logger_filters:domain/2 for a description of how this field can be used.

report_cb

If the log message is specified as a report(), the report_cb key can be associated with a fun (report callback) that converts the report to a format string and arguments, or directly to a string. See the type definition of report_cb(), and section Log Message in the User's Guide for more information about report callbacks.

-type msg_fun() :: term().

Specs

msg_fun() :: fun((term()) -> msg_fun_return() | {msg_fun_return(), metadata()}).

-type msg_fun_return() :: term().

Specs

msg_fun_return() :: {io:format(), [term()]} | report() | unicode:chardata() | ignore.

-type olp_config() :: term().

Specs

olp_config() ::
    #{sync_mode_qlen => non_neg_integer(),
      drop_mode_qlen => pos_integer(),
      flush_qlen => pos_integer(),
      burst_limit_enable => boolean(),
      burst_limit_max_count => pos_integer(),
      burst_limit_window_time => pos_integer(),
      overload_kill_enable => boolean(),
      overload_kill_qlen => pos_integer(),
      overload_kill_mem_size => pos_integer(),
      overload_kill_restart_after => non_neg_integer() | infinity}.

-type primary_config() :: term().

Specs

primary_config() ::
    #{level => level() | all | none,
      metadata => metadata(),
      filter_default => log | stop,
      filters => [{filter_id(), filter()}]}.

Primary configuration data for Logger. The following default values apply:

level => info

filter_default => log

filters => []

-type report() :: term().

Specs

report() :: map() | [{atom(), term()}].

-type report_cb() :: term().

Specs

report_cb() ::
    fun((report()) -> {io:format(), [term()]}) |
    fun((report(), report_cb_config()) -> unicode:chardata()).

A fun which converts a report() to a format string and arguments, or directly to a string. See section Log Message in the User's Guide for more information.

-type report_cb_config() :: term().

Specs

report_cb_config() ::
    #{depth := pos_integer() | unlimited,
      chars_limit := pos_integer() | unlimited,
      single_line := boolean()}.

-type timestamp() :: term().

Specs

timestamp() :: integer().

A timestamp produced with logger:timestamp().

Functions

add_handler/3
(since OTP 21.0)

Specs

add_handler(HandlerId, Module, Config) -> ok | {error, term()}
               when HandlerId :: handler_id(), Module :: module(), Config :: handler_config().

Add a handler with the given configuration.

HandlerId is a unique identifier which must be used in all subsequent calls referring to this handler.

add_handler_filter/3
(since OTP 21.0)

Specs

add_handler_filter(HandlerId, FilterId, Filter) -> ok | {error, term()}
                      when HandlerId :: handler_id(), FilterId :: filter_id(), Filter :: filter().

Add a filter to the specified handler.

The filter fun is called with the log event as the first parameter, and the specified filter_args() as the second parameter.

The return value of the fun specifies if a log event is to be discarded or forwarded to the handler callback:

log_event()

The filter passed. The next handler filter, if any, is applied. If no more filters exist for this handler, the log event is forwarded to the handler callback.

stop

The filter did not pass, and the log event is immediately discarded.

ignore

The filter has no knowledge of the log event. The next handler filter, if any, is applied. If no more filters exist for this handler, the value of the filter_default configuration parameter for the handler specifies if the log event shall be discarded or forwarded to the handler callback.

See section Filters in the User's Guide for more information about filters.

Some built-in filters exist. These are defined in logger_filters(3).

add_handlers/1
(since OTP 21.0)

Specs

add_handlers(Application) -> ok | {error, term()} when Application :: atom();
            (HandlerConfig) -> ok | {error, term()} when HandlerConfig :: [config_handler()].

Reads the application configuration parameter logger and calls add_handlers/1 with its contents.

add_handlers/1
(since OTP 21.0)

Specs

add_handlers(Application) -> ok | {error, term()} when Application :: atom();
            (HandlerConfig) -> ok | {error, term()} when HandlerConfig :: [config_handler()].

This function should be used by custom Logger handlers to make configuration consistent no matter which handler the system uses. Normal usage is to add a call to logger:add_handlers/1 just after the processes that the handler needs are started, and pass the application's logger configuration as the argument. For example:

-behaviour(application).
start(_, []) ->
    case supervisor:start_link({local, my_sup}, my_sup, []) of
        {ok, Pid} ->
            ok = logger:add_handlers(my_app),
            {ok, Pid, []};
        Error -> Error
     end.

This reads the logger configuration parameter from the my_app application and starts the configured handlers. The contents of the configuration use the same rules as the logger handler configuration.

If the handler is meant to replace the default handler, the Kernel's default handler have to be disabled before the new handler is added. A sys.config file that disables the Kernel handler and adds a custom handler could look like this:

[{kernel,
  [{logger,
    %% Disable the default Kernel handler
    [{handler, default, undefined}]}]},
 {my_app,
  [{logger,
    %% Enable this handler as the default
    [{handler, default, my_handler, #{}}]}]}].
         

add_primary_filter/2
(since OTP 21.0)

Specs

add_primary_filter(FilterId, Filter) -> ok | {error, term()}
                      when FilterId :: filter_id(), Filter :: filter().

Add a primary filter to Logger.

The filter fun is called with the log event as the first parameter, and the specified filter_args() as the second parameter.

The return value of the fun specifies if a log event is to be discarded or forwarded to the handlers:

log_event()

The filter passed. The next primary filter, if any, is applied. If no more primary filters exist, the log event is forwarded to the handler part of Logger, where handler filters are applied.

stop

The filter did not pass, and the log event is immediately discarded.

ignore

The filter has no knowledge of the log event. The next primary filter, if any, is applied. If no more primary filters exist, the value of the primary filter_default configuration parameter specifies if the log event shall be discarded or forwarded to the handler part.

See section Filters in the User's Guide for more information about filters.

Some built-in filters exist. These are defined in logger_filters(3).

alert(StringOrReport[,Metadata])
(since OTP 21.0)

Equivalent to log(alert,...).

alert(Fun,FunArgs[,Metadata]) alert(Format,Args[,Metadata]) alert(StringOrReport[,Metadata])
(since OTP 21.0)

alert(Fun,FunArgs[,Metadata]) alert(Format,Args[,Metadata])
(since OTP 21.0)

compare_levels/2
(since OTP 21.0)

Specs

compare_levels(Level1, Level2) -> eq | gt | lt
                  when Level1 :: level() | all | none, Level2 :: level() | all | none.

Compare the severity of two log levels. Returns gt if Level1 is more severe than Level2, lt if Level1 is less severe, and eq if the levels are equal.

critical(StringOrReport[,Metadata])
(since OTP 21.0)

Equivalent to log(critical,...).

critical(Fun,FunArgs[,Metadata]) critical(Format,Args[,Metadata]) critical(StringOrReport[,Metadata])
(since OTP 21.0)

critical(Fun,FunArgs[,Metadata]) critical(Format,Args[,Metadata])
(since OTP 21.0)

debug(StringOrReport[,Metadata])
(since OTP 21.0)

Equivalent to log(debug,...).

debug(Fun,FunArgs[,Metadata]) debug(Format,Args[,Metadata]) debug(StringOrReport[,Metadata])
(since OTP 21.0)

debug(Fun,FunArgs[,Metadata]) debug(Format,Args[,Metadata])
(since OTP 21.0)

emergency(StringOrReport[,Metadata])
(since OTP 21.0)

Equivalent to log(emergency,...).

emergency(Fun,FunArgs[,Metadata]) emergency(Format,Args[,Metadata]) emergency(StringOrReport[,Metadata])
(since OTP 21.0)

emergency(Fun,FunArgs[,Metadata]) emergency(Format,Args[,Metadata])
(since OTP 21.0)

error(StringOrReport[,Metadata])
(since OTP 21.0)

Equivalent to log(error,...).

error(Fun,FunArgs[,Metadata]) error(Format,Args[,Metadata]) error(StringOrReport[,Metadata])
(since OTP 21.0)

error(Fun,FunArgs[,Metadata]) error(Format,Args[,Metadata])
(since OTP 21.0)

format_report/1
(since OTP 21.0)

Specs

format_report(Report) -> FormatArgs when Report :: report(), FormatArgs :: {io:format(), [term()]}.

Convert a log message on report form to {Format, Args}. This is the default report callback used by logger_formatter(3) when no custom report callback is found. See section Log Message in the Kernel User's Guide for information about report callbacks and valid forms of log messages.

The function produces lines of Key: Value from key-value lists. Strings are printed with ~ts and other terms with ~tp.

If Report is a map, it is converted to a key-value list before formatting as such.

get_config/0
(since OTP 21.0)

Specs

get_config() ->
              #{primary => primary_config(),
                handlers => [handler_config()],
                proxy => olp_config(),
                module_levels => [{module(), level() | all | none}]}.

Look up all current Logger configuration, including primary, handler, and proxy configuration, and module level settings.

get_handler_config/0
(since OTP 21.0)

Specs

get_handler_config() -> [Config] when Config :: handler_config().

Look up the current configuration for all handlers.

get_handler_config/1
(since OTP 21.0)

Specs

get_handler_config(HandlerId) -> {ok, Config} | {error, term()}
                      when HandlerId :: handler_id(), Config :: handler_config().

Look up the current configuration for the given handler.

get_handler_ids/0
(since OTP 21.0)

Specs

get_handler_ids() -> [HandlerId] when HandlerId :: handler_id().

Look up the identities for all installed handlers.

get_module_level/0
(since OTP 21.0)

Specs

get_module_level() -> [{Module, Level}] when Module :: module(), Level :: level() | all | none.

Look up all current module levels. Returns a list containing one {Module,Level} element for each module for which the module level was previously set with set_module_level/2.

get_module_level/1
(since OTP 21.0)

Specs

get_module_level(Modules) -> [{Module, Level}]
                    when
                        Modules :: [Module] | Module,
                        Module :: module(),
                        Level :: level() | all | none.

Look up the current level for the given modules. Returns a list containing one {Module,Level} element for each of the given modules for which the module level was previously set with set_module_level/2.

get_primary_config/0
(since OTP 21.0)

Specs

get_primary_config() -> Config when Config :: primary_config().

Look up the current primary configuration for Logger.

get_process_metadata/0
(since OTP 21.0)

Specs

get_process_metadata() -> Meta | undefined when Meta :: metadata().

Retrieve data set with set_process_metadata/1 or update_process_metadata/1.

get_proxy_config/0
(since OTP 21.3)

Specs

get_proxy_config() -> Config when Config :: olp_config().

Look up the current configuration for the Logger proxy.

For more information about the proxy, see section Logger Proxy in the Kernel User's Guide.

i/0
(since OTP 21.3)

Specs

i() -> ok.

Pretty print the Logger configuration.

i/1
(since OTP 21.3)

Specs

i(What) -> ok when What :: primary | handlers | proxy | modules | handler_id().

info(StringOrReport[,Metadata])
(since OTP 21.0)

Equivalent to log(info,...).

info(Fun,FunArgs[,Metadata]) info(Format,Args[,Metadata]) info(StringOrReport[,Metadata])
(since OTP 21.0)

info(Fun,FunArgs[,Metadata]) info(Format,Args[,Metadata])
(since OTP 21.0)

log/2
(since OTP 21.0)

Specs

log(Level, StringOrReport) -> ok
       when Level :: level(), StringOrReport :: unicode:chardata() | report().

Create a log event at the given log level, with the given message to be logged and metadata. Examples:

%% A plain string
logger:log(info, "Hello World").
%% A plain string with metadata
logger:log(debug, "Hello World", #{ meta => data }).
%% A format string with arguments
logger:log(warning, "The roof is on ~ts",[Cause]).
%% A report
logger:log(warning, #{ what => roof, cause => Cause }).
        

The message and metadata can either be given directly in the arguments, or returned from a fun. Passing a fun instead of the message/metadata directly is useful in scenarios when the message/metadata is very expensive to compute. This is because the fun is only evaluted when the message/metadata is actually needed, which may be not at all if the log event is not to be logged. Examples:

%% A plain string with expensive metadata
logger:info(fun([]) -> {"Hello World", #{ meta => expensive() }} end,[]).
%% An expensive report
logger:debug(fun(What) -> #{ what => What, cause => expensive() } end,roof).
%% A plain string with expensive metadata and normal metadata
logger:debug(fun([]) -> {"Hello World", #{ meta => expensive() }} end,[],
             #{ meta => data }).
        

When metadata is given both as an argument and returned from the fun they are merged. If equal keys exists the values are taken from the metadata returned by the fun.

log/3
(since OTP 21.0)

Specs

log(Level, StringOrReport, Metadata) -> ok
       when
           Level :: level(),
           StringOrReport :: unicode:chardata() | report(),
           Metadata :: metadata();
   (Level, Format, Args) -> ok when Level :: level(), Format :: io:format(), Args :: [term()];
   (Level, Fun, FunArgs) -> ok when Level :: level(), Fun :: msg_fun(), FunArgs :: term().

log/4
(since OTP 21.0)

Specs

log(Level, Format, Args, Metadata) -> ok
       when Level :: level(), Format :: io:format(), Args :: [term()], Metadata :: metadata();
   (Level, Fun, FunArgs, Metadata) -> ok
       when Level :: level(), Fun :: msg_fun(), FunArgs :: term(), Metadata :: metadata().

notice(StringOrReport[,Metadata])
(since OTP 21.0)

Equivalent to log(notice,...).

notice(Fun,FunArgs[,Metadata]) notice(Format,Args[,Metadata]) notice(StringOrReport[,Metadata])
(since OTP 21.0)

notice(Fun,FunArgs[,Metadata]) notice(Format,Args[,Metadata])
(since OTP 21.0)

remove_handler/1
(since OTP 21.0)

Specs

remove_handler(HandlerId) -> ok | {error, term()} when HandlerId :: handler_id().

Remove the handler identified by HandlerId.

remove_handler_filter/2
(since OTP 21.0)

Specs

remove_handler_filter(HandlerId, FilterId) -> ok | {error, term()}
                         when HandlerId :: handler_id(), FilterId :: filter_id().

Remove the filter identified by FilterId from the handler identified by HandlerId.

remove_primary_filter/1
(since OTP 21.0)

Specs

remove_primary_filter(FilterId) -> ok | {error, term()} when FilterId :: filter_id().

Remove the primary filter identified by FilterId from Logger.

set_application_level/2
(since OTP 21.1)

Specs

set_application_level(Application, Level) -> ok | {error, not_loaded}
                         when Application :: atom(), Level :: level() | all | none.

Set the log level for all the modules of the specified application.

This function is a convenience function that calls logger:set_module_level/2 for each module associated with an application.

set_handler_config/2
(since OTP 21.0)

Specs

set_handler_config(HandlerId, Config) -> ok | {error, term()}
                      when HandlerId :: handler_id(), Config :: handler_config().

Set configuration data for the specified handler. This overwrites the current handler configuration.

To modify the existing configuration, use update_handler_config/2, or, if a more complex merge is needed, read the current configuration with get_handler_config/1 , then do the merge before writing the new configuration back with this function.

If a key is removed compared to the current configuration, and the key is known by Logger, the default value is used. If it is a custom key, then it is up to the handler implementation if the value is removed or a default value is inserted.

set_handler_config/3
(since OTP 21.0)

Specs

set_handler_config(HandlerId, level, Level) -> Return
                      when
                          HandlerId :: handler_id(),
                          Level :: level() | all | none,
                          Return :: ok | {error, term()};
                  (HandlerId, filter_default, FilterDefault) -> Return
                      when
                          HandlerId :: handler_id(),
                          FilterDefault :: log | stop,
                          Return :: ok | {error, term()};
                  (HandlerId, filters, Filters) -> Return
                      when
                          HandlerId :: handler_id(),
                          Filters :: [{filter_id(), filter()}],
                          Return :: ok | {error, term()};
                  (HandlerId, formatter, Formatter) -> Return
                      when
                          HandlerId :: handler_id(),
                          Formatter :: {module(), formatter_config()},
                          Return :: ok | {error, term()};
                  (HandlerId, config, Config) -> Return
                      when
                          HandlerId :: handler_id(),
                          Config :: term(),
                          Return :: ok | {error, term()}.

Add or update configuration data for the specified handler. If the given Key already exists, its associated value will be changed to the given value. If it does not exist, it will be added.

If the value is incomplete, which for example can be the case for the config key, it is up to the handler implementation how the unspecified parts are set. For all handlers in the Kernel application, unspecified data for the config key is set to default values. To update only specified data, and keep the existing configuration for the rest, use update_handler_config/3.

See the definition of the handler_config() type for more information about the different parameters.

set_module_level/2
(since OTP 21.0)

Specs

set_module_level(Modules, Level) -> ok | {error, term()}
                    when Modules :: [module()] | module(), Level :: level() | all | none.

Set the log level for the specified modules.

The log level for a module overrides the primary log level of Logger for log events originating from the module in question. Notice, however, that it does not override the level configuration for any handler.

For example: Assume that the primary log level for Logger is info, and there is one handler, h1, with level info and one handler, h2, with level debug.

With this configuration, no debug messages will be logged, since they are all stopped by the primary log level.

If the level for mymodule is now set to debug, then debug events from this module will be logged by the handler h2, but not by handler h1.

Debug events from other modules are still not logged.

To change the primary log level for Logger, use set_primary_config(level, Level).

To change the log level for a handler, use set_handler_config(HandlerId, level, Level) .

The originating module for a log event is only detected if the key mfa exists in the metadata, and is associated with {Module, Function, Arity}. When log macros are used, this association is automatically added to all log events. If an API function is called directly, without using a macro, the logging client must explicitly add this information if module levels shall have any effect.

set_primary_config/1
(since OTP 21.0)

Specs

set_primary_config(Config) -> ok | {error, term()} when Config :: primary_config().

Set primary configuration data for Logger. This overwrites the current configuration.

To modify the existing configuration, use update_primary_config/1, or, if a more complex merge is needed, read the current configuration with get_primary_config/0 , then do the merge before writing the new configuration back with this function.

If a key is removed compared to the current configuration, the default value is used.

set_primary_config/2
(since OTP 21.0,OTP 24.0)

Specs

set_primary_config(level, Level) -> ok | {error, term()} when Level :: level() | all | none;
                  (filter_default, FilterDefault) -> ok | {error, term()}
                      when FilterDefault :: log | stop;
                  (filters, Filters) -> ok | {error, term()}
                      when Filters :: [{filter_id(), filter()}];
                  (metadata, Meta) -> ok | {error, term()} when Meta :: metadata().

Add or update primary configuration data for Logger. If the given Key already exists, its associated value will be changed to the given value. If it does not exist, it will be added.

set_process_metadata/1
(since OTP 21.0)

Specs

set_process_metadata(Meta) -> ok when Meta :: metadata().

Set metadata which Logger shall automatically insert in all log events produced on the current process.

Location data produced by the log macros, and/or metadata given as argument to the log call (API function or macro), are merged with the process metadata. If the same keys occur, values from the metadata argument to the log call overwrite values from the process metadata, which in turn overwrite values from the location data.

Subsequent calls to this function overwrites previous data set. To update existing data instead of overwriting it, see update_process_metadata/1.

set_proxy_config/1
(since OTP 21.3)

Specs

set_proxy_config(Config) -> ok | {error, term()} when Config :: olp_config().

Set configuration data for the Logger proxy. This overwrites the current proxy configuration. Keys that are not specified in the Config map gets default values.

To modify the existing configuration, use update_proxy_config/1, or, if a more complex merge is needed, read the current configuration with get_proxy_config/0 , then do the merge before writing the new configuration back with this function.

For more information about the proxy, see section Logger Proxy in the Kernel User's Guide.

timestamp/0
(since OTP 21.3)

Specs

timestamp() -> timestamp().

Return a timestamp that can be inserted as the time field in the meta data for a log event. It is produced with os:system_time(microsecond).

Notice that Logger automatically inserts a timestamp in the meta data unless it already exists. This function is exported for the rare case when the timestamp must be taken at a different point in time than when the log event is issued.

unset_application_level/1
(since OTP 21.1)

Specs

unset_application_level(Application) -> ok | {error, {not_loaded, Application}}
                           when Application :: atom().

Unset the log level for all the modules of the specified application.

This function is a utility function that calls logger:unset_module_level/2 for each module associated with an application.

unset_module_level/0
(since OTP 21.0)

Specs

unset_module_level() -> ok.

Remove module specific log settings. After this, the primary log level is used for all modules.

unset_module_level/1
(since OTP 21.0)

Specs

unset_module_level(Modules) -> ok when Modules :: [module()] | module().

Remove module specific log settings. After this, the primary log level is used for the specified modules.

unset_process_metadata/0
(since OTP 21.0)

Specs

unset_process_metadata() -> ok.

Delete data set with set_process_metadata/1 or update_process_metadata/1.

update_formatter_config/2
(since OTP 21.0)

Specs

update_formatter_config(HandlerId, FormatterConfig) -> ok | {error, term()}
                           when HandlerId :: handler_id(), FormatterConfig :: formatter_config().

Update the formatter configuration for the specified handler.

The new configuration is merged with the existing formatter configuration.

To overwrite the existing configuration without any merge, use

set_handler_config(HandlerId, formatter,
	      {FormatterModule, FormatterConfig}).

update_formatter_config/3
(since OTP 21.0)

Specs

update_formatter_config(HandlerId, Key, Value) -> ok | {error, term()}
                           when HandlerId :: handler_id(), Key :: atom(), Value :: term().

Update the formatter configuration for the specified handler.

This is equivalent to

update_formatter_config(HandlerId, #{Key => Value})

update_handler_config/2
(since OTP 21.0)

Specs

update_handler_config(HandlerId, Config) -> ok | {error, term()}
                         when HandlerId :: handler_id(), Config :: handler_config().

Update configuration data for the specified handler. This function behaves as if it was implemented as follows:

{ok, {_, Old}} = logger:get_handler_config(HandlerId),
logger:set_handler_config(HandlerId, maps:merge(Old, Config)).

To overwrite the existing configuration without any merge, use set_handler_config/2 .

update_handler_config/3
(since OTP 21.2)

Specs

update_handler_config(HandlerId, level, Level) -> Return
                         when
                             HandlerId :: handler_id(),
                             Level :: level() | all | none,
                             Return :: ok | {error, term()};
                     (HandlerId, filter_default, FilterDefault) -> Return
                         when
                             HandlerId :: handler_id(),
                             FilterDefault :: log | stop,
                             Return :: ok | {error, term()};
                     (HandlerId, filters, Filters) -> Return
                         when
                             HandlerId :: handler_id(),
                             Filters :: [{filter_id(), filter()}],
                             Return :: ok | {error, term()};
                     (HandlerId, formatter, Formatter) -> Return
                         when
                             HandlerId :: handler_id(),
                             Formatter :: {module(), formatter_config()},
                             Return :: ok | {error, term()};
                     (HandlerId, config, Config) -> Return
                         when
                             HandlerId :: handler_id(),
                             Config :: term(),
                             Return :: ok | {error, term()}.

Add or update configuration data for the specified handler. If the given Key already exists, its associated value will be changed to the given value. If it does not exist, it will be added.

If the value is incomplete, which for example can be the case for the config key, it is up to the handler implementation how the unspecified parts are set. For all handlers in the Kernel application, unspecified data for the config key is not changed. To reset unspecified data to default values, use set_handler_config/3.

See the definition of the handler_config() type for more information about the different parameters.

update_primary_config/1
(since OTP 21.0)

Specs

update_primary_config(Config) -> ok | {error, term()} when Config :: primary_config().

Update primary configuration data for Logger. This function behaves as if it was implemented as follows:

Old = logger:get_primary_config(),
logger:set_primary_config(maps:merge(Old, Config)).

To overwrite the existing configuration without any merge, use set_primary_config/1 .

update_process_metadata/1
(since OTP 21.0)

Specs

update_process_metadata(Meta) -> ok when Meta :: metadata().

Set or update metadata to use when logging from current process

If process metadata exists for the current process, this function behaves as if it was implemented as follows:

logger:set_process_metadata(maps:merge(logger:get_process_metadata(), Meta)).

If no process metadata exists, the function behaves as set_process_metadata/1 .

update_proxy_config/1
(since OTP 21.3)

Specs

update_proxy_config(Config) -> ok | {error, term()} when Config :: olp_config().

Update configuration data for the Logger proxy. This function behaves as if it was implemented as follows:

Old = logger:get_proxy_config(),
logger:set_proxy_config(maps:merge(Old, Config)).

To overwrite the existing configuration without any merge, use set_proxy_config/1 .

For more information about the proxy, see section Logger Proxy in the Kernel User's Guide.

warning(StringOrReport[,Metadata])
(since OTP 21.0)

Equivalent to log(warning,...).

warning(Fun,FunArgs[,Metadata]) warning(Format,Args[,Metadata]) warning(StringOrReport[,Metadata])
(since OTP 21.0)

warning(Fun,FunArgs[,Metadata]) warning(Format,Args[,Metadata])
(since OTP 21.0)